requirements.txt
torch==0.4.1
numpy==1.14.5
matplotlib==2.1.2
tqdm==4.23.4
scipy==1.0.0
bcolz==1.2.1
easydict==1.7
opencv-python==3.4.2.17
Pillow==5.2.0
mxnet==1.2.1.post1
scikit-learn==0.19.2
tensorboardX==1.2
torchvision==0.2.1
cython==0.29.14
mxnet-cu100